feat: improved archive extraction progress reporting

+++ src/game/genshin/diff.rs
@@ -145,12 +145,12 @@ pub enum Status {
 
 pub struct Updater {
     status: Cell<Status>,
-    current: Cell<usize>,
-    total: Cell<usize>,
+    current: Cell<u64>,
+    total: Cell<u64>,
 
     worker: Option<JoinHandle<Result<(), Error>>>,
     worker_result: Option<Result<(), Error>>,
-    updater: flume::Receiver<(Status, usize, usize)>
+    updater: flume::Receiver<(Status, u64, u64)>
 }
 
 impl Updater {
@@ -207,14 +207,14 @@ impl UpdaterExt for Updater {
     }
 
     #[inline]
-    fn current(&self) -> usize {
+    fn current(&self) -> u64 {
         self.update();
 
         self.current.get()
     }
 
     #[inline]
-    fn total(&self) -> usize {
+    fn total(&self) -> u64 {
         self.update();
 
         self.total.get()
